backend/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
Flask-MongoEngine
Werkzeug
python-dotenv
